requirements.txt
django
djangorestframework
environs
mysqlclient
pymysql
django-recaptcha
requests
psycopg
djangorestframework-simplejwt
